#47 разделил api для мобильного приложения и для веб приложения. поменял роутинг
