+++ ror/turajol/app/controllers/api/api_controller.rb
@@ -1,0 +1,10 @@
+module Api
+  class ApiController < ActionController::API
+    #TODO переделать чтобы была валидация приложений, которые отправляет координаты
+    protect_from_forgery with: :null_session, if: Proc.new { |c| c.request.format == 'application/json' }
+    add_flash_types :error, :success
+  end
+end
+
+
+
