+++ ror/turajol/app/controllers/application_controller.rb
@@ -1,5 +1,3 @@
 class ApplicationController < ActionController::Base
-  #TODO переделать чтобы была валидация приложений, которые отправляет координаты
-  protect_from_forgery with: :null_session, if: Proc.new { |c| c.request.format == 'application/json' }
-  add_flash_types :error, :success
+  protect_from_forgery with: :exception
 end
