+++ ror/turajol/app/views/points/index.html.erb
@@ -3,11 +3,6 @@
       <p>
         <%= "Широта: #{point.latitude}" %>
         <%= "Долгота: #{point.longitude}" %>
-        <%= link_to "Число подтверждений: #{point.counter}", confirm_path(point) %>
-        <%= link_to "Удалить точку", point_path(point), method: :delete  %>
       </p>
   <% end %>
-</div>
-<div>
-  <%=	link_to "Добавить точку", new_point_path  %>
 </div>
