+++ ror/turajol/config/routes.rb
@@ -1,6 +1,11 @@
 Rails.application.routes.draw do
 
   root 'points#index'
-  resources :points, :only => [ :new, :create, :destroy]
-  get 'points/:id/confirm', to: 'points#confirm', as: 'confirm'
+
+  namespace :api do
+    namespace :v1 do
+      resources :points, :only => [ :index, :new, :create]
+      get 'points/:id/confirm', to: 'points#confirm', as: 'confirm'
+    end
+  end
 end
